Incorporate git commit hook into repository...
This feature is no longer experimental, although further development
will follow.

The pre-commit-Lucy binary may be used as a commit hook for git.
It works by running Lucy on all added or modified .lua files when
a new commit is being made. If Lucy finds something to complain about,
it writes to stderr and exits with non-zero exit code. This code
is propagated by commit hook, aborting the commit action. Lucy's error
messages are printed out on stdout.

Remember that you may easily skip the pre-commit hook with --no-verify
option.